Jframe settings has been moved to the Board class
