+++ src/com/beltorion/wanderer/repositories/GameMap.java
@@ -1,0 +1,2 @@
+package com.beltorion.wanderer.repositories;public class GameMap {
+}
